// src/contract/model/bookingBriefDto.ts
/**
 * EV Rental API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface BookingBriefDto {
  startDate?: string;
  endDate?: string;
  renterId?: string;
  bookingId?: string;
}